requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
crispy-bootstrap3==2022.1
crispy-bootstrap4==2022.1
Django==4.1.7
django-crispy-forms==2.0
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
freezegun==1.2.2
hiredis==2.2.3
iniconfig==2.0.0
Markdown==3.4.1
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.5
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
